게시판 수정 삭제 권한부여

+++ src/main/java/com/example/travelback/board/controller/Boardcontroller.java
@@ -3,7 +3,9 @@ package com.example.travelback.board.controller;
 
 import com.example.travelback.board.domain.Board;
 import com.example.travelback.board.service.BoardService;
+import com.example.travelback.user.dto.Member;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,8 +30,15 @@ public class Boardcontroller {
 
 
     @PostMapping("add")
-    public ResponseEntity add(@RequestBody Board board){
+    public ResponseEntity add(@RequestBody Board board,
+                              @SessionAttribute(value = "login" ,required = false)Member login){
         System.out.println("Board = " + board);
+        System.out.println("login = " + login);
+
+        // 로그인 여부확인
+        if(login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
         // 검증하기
         if (!service.validate(board)){
@@ -37,7 +46,7 @@ public class Boardcontroller {
         }
 
         // 글 추가 하기
-        if (service.add(board)){
+        if (service.add(board,login)){
         return     ResponseEntity.ok().build();
         }
         else {
@@ -60,7 +69,19 @@ public class Boardcontroller {
 
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id){
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute(value = "login",required = false)Member login){
+
+        //401 권한 없음 에러
+        if (login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        //403 서버에서 거절
+        if (!service.hasAccess(id,login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+
          if (service.remove(id)){
              return ResponseEntity.ok().build();
          }else {
@@ -69,11 +90,21 @@ public class Boardcontroller {
     }
 
     @PutMapping("edit")
-    public  void update(
-            @RequestBody Board board){
-        System.out.println("board = " + board);
+    public  ResponseEntity update(
+            @RequestBody Board board, @SessionAttribute (value = "login",required = false)Member login){
 
-        service.update(board);
+        if (login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.hasAccess(board.getId(),login)){
+            return  ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        if (service.update(board)) {
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.internalServerError().build();
+        }
 
     }
